requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.8.1
attrs==17.4.0
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
bitstring==3.1.7
bluepy==1.3.0
configparser==4.0.2
constants==0.6.0
contextlib2==0.6.0.post1
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.21
decorator==4.1.2
distlib==0.3.1
enum34==1.1.10
filelock==3.0.12
funcsigs==1.0.2
futures==3.3.0
gast==0.4.0
gps==3.17
graphsurgeon==0.4.1
grpcio==1.34.0
idna==2.6
imbalanced-learn==0.4.3
importlib-metadata==2.1.1
importlib-resources==3.3.1
ipaddress==1.0.17
Jetson.GPIO==2.0.4
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.5
mne==0.15
mock==3.0.5
nose==1.3.7
numpy==1.14.5
olefile==0.45.1
# Editable install with no version control (OpenBCI-Python==1.0.0)
-e /home/johnny/.local/lib/python2.7/site-packages
pandas==0.24.2
pathlib2==2.3.5
pbr==5.5.1
Pillow==5.1.0
pluggy==0.6.0
protobuf==3.14.0
py==1.5.2
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pylsl==1.12.2
pyOpenBCI==0.13
pyparsing==2.4.7
pyserial==2.7
pytest==3.3.2
python-dateutil==2.8.1
python-osc==1.6.3
pytz==2021.1
pyxdg==0.25
requests==2.7.0
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.3
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.10.0
socketIO-client==0.6.5
some-package==0.1
stevedore==1.32.0
subprocess32==3.5.4
tensorboard==1.10.0
tensorflow-gpu==1.10.0+nv18.9
tensorrt==6.0.1.10
termcolor==1.1.0
typing==3.7.4.3
uff==0.6.5
unity-lens-photos==1.0
urwid==2.0.1
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
websocket-client==0.32.0
Werkzeug==1.0.1
xmltodict==0.12.0
Yapsy==1.11.23
zipp==1.2.0
